Update patch set 2
Patch Set 2:

For a better tracking, why don't you file a bug at v8 bug tracker and refer to it here in Bug: line?  

Is there any test in test262/ecma402 that can pass now?  If so, you may as well want to update test262/test262.status and mention them in Test= line in the CL description.

Patch-set: 2