Update patch set 1
Patch Set 1: Code-Review+1

LGTM.

Note that one of the considerations when we chose the serialization format was that it was 32/64 bit agnostic, to make it possible to read data persisted to disk when upgrading from 32-bit to 64-bit Chrome. I guess that use case does not apply to big endian embedders -- but if you do want to support it, then a better fix would be to have the serialized format be the same (little-endian) everywhere, and adapt the serialization/deserialization code for big endian platforms to perform the endianness conversion.

Patch-set: 1
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1