ThreadDataTable: Change linked list to hash map.
For use cases with a large number of threads or a large number of isolates (or
both), ThreadDataTable can be a major performance bottleneck due to O(n)
lookup time of the linked list. Switching to a hash map reduces this to O(1).

Example 1: Sandstorm.io, a Node.js app that utilizes "fibers", was observed
spending the majority of CPU time iterating over the ThreadDataTable.
See: https://sandstorm.io/news/2016-09-30-fiber-bomb-debugging-story

Example 2: Cloudflare's Workers engine, a high-multi-tenancy web server
framework built on V8 (but not Node), creates large numbers of threads and
isolates per-process. It saw a 34x improvement in throughput when we applied
this patch.

This commit is exactly the patch which has been used in the Workers engine since
launch. It is based on an earlier patch developed by Drew Fisher at Sandstorm.
(I, Kenton Varda, am both the co-founder/CEO of Sandstorm and the tech lead of
Cloudflare Workers.)

Bug: v8:5338
Change-Id: If0d11509afb2e043b888c376e36d3463db931b47