[prof] Attribute ticks in embedded builtins correctly
When attributing ticks to a function, we first check if the current pc
matches a section within the dynamic library. If we find a match here,
then we don't continue looking within dynamically generated range
information, e.g. for JS functions and builtins.

This logic breaks when embedded builtins come into play. They live
within a shared library, and are found when looking up statics. But
what we really want is to look up the dynamically generated
code-range, which contains more precise information.

Bug: v8:6666
Change-Id: I7cea2cd4898f5a08381a071bdbc2f862b9c80880