Allow use of ordinary shared array buffer in wasm threads
For wasm memory buffers, we normally require the memory to be allocated
via WebAssembly.Memory, which will set a is_wasm_memory flag and
register the memory with the wasm allocation tracker. This CL weakens
that requirement in a DCHECK to allow for running experimental threaded
applications even though the is_wasm_memory flag is not currently
propagated via postMessage.

Bug: 836800
Change-Id: I4613b8651423307ce4cd466c0df28fc43244ec4f