[heap] Fine-tune write barrier for strong references
By specializing RecordWrite and InNewSpace/InToSpace/InFromSpace
for references that are known not to be weak, we can avoid a few
unnecessary instructions. The performance difference is probably
not measurable, but at least this reduces code size: it replaces
the sequence
  and ecx, 0xfffffffd  ; mask out "weak" bit
  dec ecx              ; remove heap object tag
  and ecx, 0xffff8000  ; get page
with a single "and" instruction to get the page directly.

Change-Id: Ic9328d7d6f5c01073288a3e87931ea6095750740