Update patch set 2
Patch Set 2:

This is not safe to merge back because it depends on the change that actually makes instances reachable from stack frames. Hence it depends on the following CL, which is not part of the 6.7 branch:

https://chromium-review.googlesource.com/c/v8/v8/+/1033913

Can you elaborate on why you merged this to the 6.7 branch?

Patch-set: 2
Reviewer: Gerrit User 1118575 <1118575@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>