Update patch set 4
Patch Set 4:

(2 comments)

This API change breaks ABI compatibility. We want to update Node.js 10 LTS to V8 6.8 eventually, which means 6.8 needs to be ABI compatible with 6.7.

That means that in this current state, when we try to update Node.js 10 to 6.8, we would have to revert this API change so that all new side-effect features you added since 6.7 has branched will not be usable in Node.js 10.

The alternative is to create a new method with the additional argument instead of using optional arguments. After 6.8 branch we can then change it back to optional arguments. This applies to a few other APIs related to side-effect flags that you landed recently.

Patch-set: 4
Reviewer: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>