Update patch set 6
Patch Set 6:

(2 comments)

> This API change breaks ABI compatibility. We want to update Node.js 10 LTS to V8 6.8 eventually, which means 6.8 needs to be ABI compatible with 6.7.

Thanks for the heads-up.  Is there a timeframe for the upgrade, or an estimate of how long these side-effect changes would stay reverted?

> The alternative is to create a new method with the additional argument instead of using optional arguments. After 6.8 branch we can then change it back to optional arguments. This applies to a few other APIs related to side-effect flags that you landed recently.

I can create a new method in this CL, and address the other side-effect flags in a separate one.  Based on this crbug [1], it seems like adding a method is ABI safe and changing signatures is not.  One of my old CLs adds a new entry to `enum class PropertyHandlerFlags`, which seems like it needs to be reverted, but is there some way to double check?

Patch-set: 6
Reviewer: Gerrit User 1230559 <1230559@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>