Merged: [regexp] Do not assume fast regexp results are non-empty
This is a backmerge of
https://chromium-review.googlesource.com/c/v8/v8/+/1061455.

It is possible for user code to modify fast regexp result objects
before they are used e.g. by RegExp.p.match, so we may not make any
assumptions about their contents. The only exception is when the
RegExp itself is fast.

No-Try: true
No-Presubmit: true
No-Treechecks: true
Bug: chromium:843022
Change-Id: Ifab9b29e21b8516926bcef3d78396fe42b9f8669