Update patch set 1
Patch Set 1:

This is a great start, thanks Maya! 

One high level question I have is: should we be testing the async hooks API? Or should we test the PromiseHooks API that's specified as part of the V8 API?

For example, we should be testing that the:
  (a) the hook was correctly called
  (b) correct promise objects were passed to the callback. 

This API does test parts of these, but not directly.

Patch-set: 1
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>