[wasm] Check representation for secondary parameter locations
A previous CL (https://chromium-review.googlesource.com/c/v8/v8/+/1075056)
introduced an optimization the spill slot where the WASM instance
is stored using the "secondary parameter location" mechanism used for
JS functions and contexts. However the optimization checked the full
machine type of the parameter, which was too narrow. As a result,
the optimization never activated. This CL fixes that by only
checking the machine representation.

R=mstarzinger@chromium.org

Change-Id: I60813935c8d119d2ddf794c797dad314b99ea867