Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> Note that the debugger already supports a stepping over yields and async calls (see v8::internal::Debug::ThreadLocal::suspended_generator_). I wonder whether that should be changed/reimplemented to fit into your model of separating responsibilities?

I discussed this with Dmitry a little more offline. I believe that current status quo is good enough. Since step-next is not only for async functions but for generators as well and actually mostly for generators. In this case V8 has enough information to make step-next by itself and does not need inspector help.

step-out for async functions is different, V8 by itself does not have enough information to make this kind of step and we on inspector side can help.

Actually I can imagine more trickier case:
- imagine we have step-in fetch call, fetch goes to service worker and pause there, and fetch call is awaited on page side,
- user click step-out inside service worker code and expect to go back to the page.

In this case v8 and inspector do not have enough information since service worker and page live in different worlds and it means that to implement step-out in this case we may end up with having some logic on DevTools frontend side where we have enough information. :)

Patch-set: 9