Add schedulers for all builders on main console
Most of this CL was generated with script, which was updated to also generate
parts of luci-scheduler.cfg configs: https://crrev.com/c/1010350. One important
assumption that the new version makes is that any builder that is defined, but
not scheduled by one of the schedulers in the master config is automatically
considered as triggered by another builder, which grants v8-ci-builder@ service
account permission to schedule it.

R=machenbach@chromium.org

No-Tree-Checks: true
Bug: chromium:793687
Change-Id: If0bb63be433c2b62d47c4118ad4388ad0e798914