Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @anna: this change means that the initial atomics wait callback occurs while the mutex is locked. It doesn't seem like that would be a problem, but I wanted to make sure this works with the node embedding.

I’ll try to check, but initially this was indeed a problem for me because it was messing with mutex locking order, and the reason for the original patch that limited the critical section scope. If I remember correctly, it was something like this:

- The initial wait callback was called with V8’s FutexEmulation::mutex_ locked, but accessed embedder information about the waiting threads that was protected through a separate mutex (i.e. lock order was V8 mutex < embedder mutex)
- Another thread could wake up the waiting thread later. It had to hold the embedder mutex while doing so to protect the same embedder data structures; but waking up the waiting thread meant indirectly calling `FutexWaitListNode::NotifyWake()`, which in turn would acquire the V8 mutex (i.e. lock order was embedder mutex < V8 mutex).

That was picked up by concurrency correctness checking tooling as something that could potentially cause a deadlock – I didn’t reproduce one, but I think it could be a real issue.

I do see the issue that is being fixed here, though. Can we maybe move only the `if (*p != value)` check and the `node->...` setup into the critical section instead? That would mean a slight API change for the callback, but I think that’s preferable here…

(Happy to volunteer code if it helps)

Patch-set: 1