Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Hi Rodrigo, did you see https://cs.chromium.org/chromium/src/v8/src/heap/spaces.h?sq=package:chromium&g=0&l=1000

We discussed this offline with Rodrigo and concluded that we need a design-doc that describes the two possible approaches: accounting in the marker and accounting in the sweeper (a.k.a array-buffer-tracker). This will help with making the decision.

Patch-set: 3
Reviewer: Gerrit User 1002405 <1002405@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>