Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> This has caused certain test files, e.e. test/debugger/debug/debug-sourceinfo.js, to be checked out with CRLF and corresponding tests to start failing, see https://crbug.com/v8/7775. Interestingly, on Buildbot these files were already checked out before this landed and continued to contain LF line endings since we re-use checkout between builds. As a result we  didn't notice this issue until trying to migrate these bots to LUCI, where a fresh checkout was created and those tests now fail.
> 
> Therefore we need to either
>  a) revert this CL
>  b) restrict which files these attributes apply to
>  c) try to find a different attribute that would still convert files to LF on check-in, but will continue to check out files with LF line endings on Windows
> 
> Interesting reads:
>  - https://git-scm.com/docs/gitattributes
>  - https://stackoverflow.com/questions/45277143/gitattributes-does-text-auto-still-uses-core-autocrlf-to-detect-eol

I am not sure if I fully get it. With "text=auto", according to the documentation the line endings of newly checked-in files is automatically converted to LF. It should never convert anything to CRLF. Also, it should only affect files which are added to the repository, not existing ones (but I just checked, all files in the v8 repo seem to be converted to LF correctly).
Do I read the documentation wrong?

Patch-set: 4