Update patch set 6
Patch Set 6:

Joyee, this change somehow broke a test [0] on the gc-stress custom snapshot builder. What that builder does is to create a custom snapshot that embeds test/mjsunit/mjsunit.js. This CL probably also changes the snapshot and that must have triggered something that somehow affected optimization decisions?

In any case, it would be great to understand this. It might not be a fundamental problem in this change, but we have a fast-revert policy in order to not block rolling into Chrome.

[0] https://cs.chromium.org/chromium/src/v8/test/mjsunit/compiler/native-context-specialization-hole-check.js?q=native-contex&sq=package:chromium&g=0&l=1

Patch-set: 6
CC: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>