[parser] propagate async arrow formal parameter error in arrow parameters
The fix is written the way it is because:

1) want to make it concise, but existing strategies for retrieving
classifier errors require the error to exist, so added a way to get
a Maybe<ExpressionClassifier::Error> which doesn't impose this
requirement.

2) The method takes a lambda, because I felt that names like
`GetErrorIfPreviousClassifierDoesntHaveThisError` or something looked
pretty bad. But if someone has a nicer name to use, then by all means
lets do that instead of using lambdas.

BUG=v8:7817
R=gsathya@chromium.org, adamk@chromium.org, littledan@chromium.org

Cq-Include-Trybots: luci.v8.try:v8_linux_noi18n_rel_ng
Change-Id: I0a84a25d7d42f89200b6fbf05eab3d39ab51fb10