[GetIsolate] Remove GetHeap in object setters
Remove the GetHeap call in object setters, replacing it with a new
Heap::FromHeapObject method which explicitly tests for the object
not being in RO space. The hope is that this method is used sparingly.

As a drive-by, memoize the heap access in the write barrier macros,
to avoid having to calculate the heap location twice.

Bug: v8:7786
Change-Id: I26d1c41c673eddced17be316da6d51a16ac853e7