Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I am not sure if this is correct. As I mentioned in my email earlier, lowering is supposed to emulate SIMD instructions. So, I would guess, there should be no difference between big endian and little endian implementations except in load and store (the difference in layout for individual data like int32 should be handled by code-gen for the corresponding nodes).

Hello Aseem,
I'm still investigating everything that has to do with BE specifics in simd-scalar-lowering. Since there are lot of different approaches in handling BE in tests and in instructions, it's hard to recognize expected behaviour of code. I cannot give an estimate on when this task will be completed and ready for upload.
Since these tests are failing right now on mips BE, maybe we should upload this fix, just for now, until I manage to take care of this problem, or skip these tests in case of big endian? What are you thoughts on this?

Patch-set: 1
Reviewer: Gerrit User 1211945 <1211945@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>