[torque] Fix crash when calling builtin via function pointer
When calling a builtin via a function pointer, torque tries to find any
builtin with the same parameter types for a descriptor. If no such
builtin exist, we currently crash.

This can happen when type aliases are used, that get the signature
wrong, and the function pointer themselves are retrieved via a
"unsafe_cast".

R=tebbi@chromium.org

Bug: v8:7793
Change-Id: Ia7436dc6541aca5de2e8dcb6b2a09978a1af9d39