[modules] Make exported variable indices stable
A module's exported variables are assigned a cell index based on their
order in the exported variable map. This map is keyed by the variable
name, an AstRawString*.

Unfortunately, these string pointers are not guaranteed to increase
monotonically as variables are encountered, which means that this map
isn't stable across parses. In particular, it can cause failures for
setVariableValue if the parser is unlucky.

This patch adds a custom comparator to these AstRawString* keyed maps,
which is stable across parses.

Change-Id: Ie6e88fc2d252d873de661d7fc5278feba3955727